+++ tests/unit/beacons/test_adb.py
