+++ tests/unit/beacons/test_avahi_announce.py
