+++ tests/unit/beacons/test_bonjour_announce.py
