+++ tests/unit/beacons/test_diskusage.py
