+++ tests/unit/beacons/test_glxinfo.py
