+++ tests/unit/beacons/test_haproxy.py
