+++ tests/unit/beacons/test_inotify.py
