+++ tests/unit/beacons/test_journald.py
