+++ tests/unit/beacons/test_load.py
