+++ tests/unit/beacons/test_log.py
