+++ tests/unit/beacons/test_memusage.py
