+++ tests/unit/beacons/test_network_info.py
