+++ tests/unit/beacons/test_ps.py
