+++ tests/unit/beacons/test_salt_proxy.py
