+++ tests/unit/beacons/test_sensehat.py
