+++ tests/unit/beacons/test_service.py
