+++ tests/unit/beacons/test_status.py
