+++ tests/unit/beacons/test_telegram_bot_msg.py
